# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Management API",
        "description": "API for managing library operations including books, users, and borrowing.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/borrows": {
            "get": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Get list of borrows",
                "operationId": "68be1687c1c822f1907d8a6b5094a562",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "borrowed",
                                "overdue",
                                "returned",
                                "cancelled",
                                "return_requested",
                                "return_rejected"
                            ]
                        }
                    },
                    {
                        "name": "book_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrow records retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Request a borrow (Member action)",
                "operationId": "06257fab65385cdb55c976d5ff5d53d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "I would like to borrow this book."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Borrow record created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/borrows/{borrow}": {
            "get": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Get borrow details",
                "operationId": "434fbf8e0693c4da9937dc7f00e51650",
                "parameters": [
                    {
                        "name": "borrow",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrow record retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Borrow not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Approve a borrow (Admin action)",
                "operationId": "c10d01bcac1a3c850309b60a733e91dd",
                "parameters": [
                    {
                        "name": "borrow",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-23"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Borrow record updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Borrow not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Delete a borrow record (Admin action)",
                "operationId": "c2daf0fa54ce27a38d6e3e2fe3f2c022",
                "parameters": [
                    {
                        "name": "borrow",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Borrow record deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Borrow not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/borrows/{borrow}/reject": {
            "put": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Reject & Cancel a borrow",
                "operationId": "9e5f5b017858bbb894348fd1b35e50bd",
                "parameters": [
                    {
                        "name": "borrow",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Rejected due to some reason."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Borrow request rejected successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Borrow not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/borrows/overdue": {
            "get": {
                "tags": [
                    "Borrows"
                ],
                "summary": "Get overdue borrows",
                "operationId": "a1e068b9294ef7d43bd4957ad8193710",
                "responses": {
                    "200": {
                        "description": "Overdue borrow records retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/returns/{borrow}/request": {
            "post": {
                "tags": [
                    "Returns"
                ],
                "summary": "Request to return a book",
                "operationId": "94f5f3861713e16c0ff0c8d2194241df",
                "parameters": [
                    {
                        "name": "borrow",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Returning the book."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return request submitted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Borrow not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/returns/{borrow}/approve": {
            "put": {
                "tags": [
                    "Returns"
                ],
                "summary": "Approve a book return (Admin action)",
                "operationId": "b28b2b7f878b607edffc581e992e4c79",
                "parameters": [
                    {
                        "name": "borrow",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Return approved."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return approved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Borrow not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/returns/{borrow}/reject": {
            "put": {
                "tags": [
                    "Returns"
                ],
                "summary": "Reject a book return (Admin action)",
                "operationId": "0b4ffb0bacf95b87925b20f34441fc19",
                "parameters": [
                    {
                        "name": "borrow",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Return rejected."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return rejected successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Borrow not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}